chore: move maintenance in a single file
